+++ extensions/iridium/src/components/NotInstalled.tsx
@@ -1,0 +1,6 @@
+import { Detail } from "@raycast/api";
+import { DownloadIridiumText } from "../constants";
+
+export function NotInstalled() {
+  return <Detail markdown={DownloadIridiumText} />;
+}