+++ extensions/iridium/src/components/index.ts
@@ -1,13 +1,2 @@
-import { HistoryItemActions, NewTabActions, TabListItemActions } from "./actions";
-import { HistoryItem, TabListItem } from "./list";
-
-export class IridiumActions {
-  public static NewTab = NewTabActions;
-  public static TabList = TabListItemActions;
-  public static TabHistory = HistoryItemActions;
-}
-
-export class IridiumListsItems {
-  public static TabList = TabListItem;
-  public static TabHistory = HistoryItem;
-}
+export * from "./IridiumActions";
+export * from "./IridiumListsItems";