+++ extensions/iridium/src/constants.ts
@@ -1,0 +1,15 @@
+export const defaultIridiumProfilePath = ["Application Support", "Iridium", "Default"];
+
+export const DownloadIridiumText = `
+  # 🚨Error: Vivaldi browser is not installed
+  ## This extension depends on Vivaldi browser. You must install it to continue.
+  
+  If you have [Homebrew](https://brew.sh/) installed then press ⏎ (Enter Key) to install Vivaldi browser.
+  [Click here](https://vivaldi.com/download/) if you want to download manually.
+  
+  [![Vivaldi](https://vivaldi.com/wp-content/themes/vivaldicom-theme/img/press/logos/vivaldi_logo_dark.png)]()
+`;
+
+export const DEFAULT_ERROR_TITLE = "An Error Occurred";
+
+export const IRIDIUM_NOT_INSTALLED_MESSAGE = "Iridium not installed";